+++ NationFlags/Views/GameViews/Cities/CityToNameView.swift
@@ -272,10 +272,10 @@ struct CityToNameView: View {
         .alert(isPresented: $settings.finish) {
             Alert(title: Text("Game over!"), message: Text("score <\(self.score)>"),
                   primaryButton: .destructive(Text("Back")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "CityToName")
+                    self.settings.saveScore(score: self.score, view: "CityToName")
                     self.presentationMode.wrappedValue.dismiss()},
                   secondaryButton: .default(Text("SaveAndPlay")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "CityToName")
+                    self.settings.saveScore(score: self.score, view: "CityToName")
                     self.playAgain()
                 })
         }
