+++ NationFlags/Views/GameViews/Cities/NameToCityErrorView.swift
@@ -242,10 +242,10 @@ struct NameToCityErrorView: View {
         .alert(isPresented: $settings.finish) {
             Alert(title: Text("Game over!"), message: Text("score <\(self.score)>"),
                   primaryButton: .destructive(Text("Back")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "NameToCityError")
+                    self.settings.saveScore(score: self.score, view: "NameToCityError")
                     self.presentationMode.wrappedValue.dismiss()},
                   secondaryButton: .default(Text("SaveAndPlay")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "NameToCityError")
+                    self.settings.saveScore(score: self.score, view: "NameToCityError")
                     self.playAgain()
                 })
         }
