+++ NationFlags/Views/GameViews/Cities/NameToCityView.swift
@@ -272,10 +272,10 @@ struct NameToCityView: View {
         .alert(isPresented: $settings.finish) {
             Alert(title: Text("Game over!"), message: Text("score <\(self.score)>"),
                   primaryButton: .destructive(Text("Back")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "NameToCity")
+                    self.settings.saveScore(score: self.score, view: "NameToCity")
                     self.presentationMode.wrappedValue.dismiss()},
                   secondaryButton: .default(Text("SaveAndPlay")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "NameToCity")
+                    self.settings.saveScore(score: self.score, view: "NameToCity")
                     self.playAgain()
                 })
         }
