+++ NationFlags/Views/GameViews/Flags/FlagToNameErrorView.swift
@@ -213,10 +213,10 @@ struct FlagToNameErrorView: View {
         .alert(isPresented: $settings.finish) {
             Alert(title: Text("Game over!"), message: Text("score <\(self.score)>"),
                   primaryButton: .destructive(Text("Back")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "FlagToNameError")
+                    self.settings.saveScore(score: self.score, view: "FlagToNameError")
                     self.presentationMode.wrappedValue.dismiss()},
                   secondaryButton: .default(Text("SaveAndPlay")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "FlagToNameError")
+                    self.settings.saveScore(score: self.score, view: "FlagToNameError")
                     self.playAgain()
                 })
         }
