+++ NationFlags/Views/GameViews/Flags/FlagToNameView.swift
@@ -214,10 +214,10 @@ struct FlagToNameView: View {
         .alert(isPresented: $settings.finish) {
             Alert(title: Text("Game over!"), message: Text("score <\(self.score)>"),
                   primaryButton: .destructive(Text("Back")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "FlagToName")
+                    self.settings.saveScore(score: self.score, view: "FlagToName")
                     self.presentationMode.wrappedValue.dismiss()},
                   secondaryButton: .default(Text("SaveAndPlay")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "FlagToName")
+                    self.settings.saveScore(score: self.score, view: "FlagToName")
                     self.playAgain()
                 })
         }
