+++ NationFlags/Views/GameViews/Flags/NameToFlagErrorView.swift
@@ -37,7 +37,9 @@ struct NameToFlagErrorView: View {
     var body: some View {
         VStack{
             Spacer()
-            FlagImage(image: settings.correctOption?.image).frame(minWidth: 0, maxWidth: .infinity, minHeight: 0, maxHeight: 300, alignment: .center).padding(.horizontal)
+            FlagImage(image: settings.correctOption?.image)
+                .frame(minWidth: 0, maxWidth: .infinity, minHeight: 0, maxHeight: 300, alignment: .center)
+                .padding(.horizontal)
             if (settings.options.count > 0) {
                 if settings.options[0].name == settings.correctOption?.name {
                     if self.disableAll {
@@ -232,10 +234,10 @@ struct NameToFlagErrorView: View {
         .alert(isPresented: $settings.finish) {
             Alert(title: Text("Game over!"), message: Text("score <\(self.score)>"),
                   primaryButton: .destructive(Text("Back")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "NameToFlagError")
+                    self.settings.saveScore(score: self.score, view: "NameToFlagError")
                     self.presentationMode.wrappedValue.dismiss()},
                   secondaryButton: .default(Text("SaveAndPlay")) {
-                    self.settings.saveScore(score: Int64(self.score), view: "NameToFlagError")
+                    self.settings.saveScore(score: self.score, view: "NameToFlagError")
                     self.playAgain()
                 })
         }
